server/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
python-dotenv
dj-database_url
psycopg2-binary
gunicorn
whitenoise
Pillow
razorpay
requests
geopy
djangorestframework-simplejwt
google-auth
cloudinary
django-cloudinary-storage
